Move DrawMap and components to their own folder
This seems like a better organiztional structure for these components.
The index file allows import statements for DrawMap to remain
at the top level of the components directory while still grouping
the main and auxillary components all in the same folder.
