+++ src/app/src/components/DrawMap/DeletePolygonConfirmModal.js
