+++ src/app/src/components/DrawMap/EditingPolygon.js
