+++ src/app/src/components/DrawMap/index.js
@@ -1,0 +1,1 @@
+export { default } from './DrawMap';
